src/it/uniroma3/diadia/diadia/Partita.java
package it.uniroma3.diadia.diadia;
import it.uniroma3.diadia.ambienti.Labirinto;
import it.uniroma3.diadia.ambienti.Stanza;
import it.uniroma3.diadia.giocatore.Giocatore;

/**
 * Questa classe modella una partita del gioco
 *
 * @author  docente di POO
 * @see Stanza
 * @version base
 */

public class Partita {

	private Labirinto labirinto;
	private Giocatore giocatore;
	private boolean finita;
	
	public Partita(){
		labirinto = new Labirinto();
		this.finita = false;
		giocatore = new Giocatore();
		labirinto.creaStanze();
	}
	
	public Labirinto getLabirinto() {
		return labirinto;
	}



	public void setLabirinto(Labirinto labirinto) {
		this.labirinto = labirinto;
	}
	
	/**
	 * Restituisce vero se e solo se la partita e' stata vinta
	 * @return vero se partita vinta
	 */
	public boolean vinta() {
		return this.labirinto.getStanzaCorrente().equals(this.labirinto.getStanzaVincente());
	}

	/**
	 * Restituisce vero se e solo se la partita e' finita
	 * @return vero se partita finita
	 */
	public boolean isFinita() {
		return finita || vinta() || (this.giocatore.getCfu() == 0);
	}

	/**
	 * Imposta la partita come finita
	 *
	 */
	public void setFinita() {
		this.finita = true;
	}



	public Giocatore getGiocatore() {
		return this.giocatore;
	}



	public void setGiocatore(Giocatore giocatore) {
		this.giocatore = giocatore;
	}




}
